# perf: don't use `#`