I haven't found a source that says it explicitly, but adding `#` to private properties will significantly degrade performance. "Hello World" was slower than `fetch` and itty-router, so I thought it was strange, investigated, and found it.

The benchmark results:

Node.js:

```
cpu: Apple M1 Pro
runtime: node v20.10.0 (arm64-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
noop                 325.93 ps/iter     (300 ps … 11.16 ns)    325 ps  345.8 ps  354.2 ps

• Get /hello
----------------------------------------------------------- -----------------------------
fetch                  4.43 µs/iter     (3.58 µs … 1.04 ms)      4 µs   8.96 µs  11.38 µs
Hono                   5.48 µs/iter     (5.29 µs … 6.22 µs)   5.52 µs   6.22 µs   6.22 µs
Hono without sharps    4.17 µs/iter     (4.03 µs … 4.45 µs)   4.25 µs   4.45 µs   4.45 µs
itty-router            4.73 µs/iter      (4.6 µs … 4.92 µs)   4.76 µs   4.92 µs   4.92 µs

summary for Get /hello
  Hono without sharps
   1.06x faster than fetch
   1.13x faster than itty-router
   1.31x faster than Hono
```

Bun:
```
cpu: Apple M1 Pro
runtime: bun 1.0.25 (arm64-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
noop                 327.07 ps/iter   (304.1 ps … 12.31 ns)    325 ps    350 ps  358.4 ps

• Get /hello
----------------------------------------------------------- -----------------------------
fetch                573.14 ns/iter (522.54 ns … 747.17 ns) 580.23 ns 747.17 ns 747.17 ns
Hono                   1.06 µs/iter   (680.44 ns … 1.68 µs)   1.32 µs   1.68 µs   1.68 µs
Hono without sharps  492.24 ns/iter (430.88 ns … 600.82 ns) 518.88 ns 569.58 ns 600.82 ns
itty-router            1.17 µs/iter     (1.05 µs … 1.41 µs)   1.21 µs   1.41 µs   1.41 µs

summary for Get /hello
  Hono without sharps
   1.16x faster than fetch
   2.15x faster than Hono
   2.37x faster than itty-router
```

The benchmark script: https://github.com/yusukebe/cloudflare-workshop-examples/tree/main/projects/benchmarks

### Author should do the followings, if applicable

- [ ] Add tests
- [x] Run tests
- [x] `yarn denoify` to generate files for Deno