@yusukebe 2024-03-03
Hey @usualoma !

What do you think? It's interesting results.