@usualoma 2024-03-03
I see - I don't think there is any difference in the following benchmarks, for example, so I'm curious where on earth such a difference would appear.
I agree with the merging, but I would like to clarify the cause before doing so.

```ts
import { run, bench, group } from "mitata";

class A {
  private x: number;
  constructor(x: number) {
    this.x = x;
  }
  public inc() {
    this.x++;
  }
}

class B {
  #x: number;
  constructor(x: number) {
    this.#x = x;
  }
  public inc() {
    this.#x++;
  }
}

bench("noop", () => {});

group({ name: "privete or hard private?" }, () => {
  bench("private", async () => {
    const a = new A(0);
    a.inc();
  });

  bench("#", async () => {
    const b = new B(0);
    b.inc();
  });
});

await run();
```