@yusukebe 2024-03-03
@usualoma 


Ahhh, I was writing the same thing!

---

You are right, there was no difference when I ran the code I wrote in TypeScript with the `tsx` or `bun` commands. However, once I run `tsc` and write it out to a JavaScript file, the result is different.

Command:

```
$ tsc
$ node ./dist/benchmark.mjs
```

Result:

```
cpu: Apple M1 Pro
runtime: node v20.10.0 (arm64-darwin)

benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
noop           63 ps/iter        (20 ps … 230 ns)     61 ps     82 ps    122 ps !

• privete or hard private?
------------------------------------------------- -----------------------------
private     49.31 ns/iter      (44.9 ns … 217 ns)  46.73 ns    119 ns    140 ns
#             233 ns/iter     (121 ns … 2'725 ns)    177 ns  1'959 ns  2'492 ns

summary for privete or hard private?
  private
   4.73x faster than #
```

This means that the code generated is different for hard private and soft private. In this case, I think the hard priavte is slower because it creates an instance of `WeakMap` and the method `__classPrivateField*` is executed.

Generated JS code:

```js
var __classPrivateFieldSet = (this && this.__classPrivateFieldSet) || function (receiver, state, value, kind, f) {
    if (kind === "m") throw new TypeError("Private method is not writable");
    if (kind === "a" && !f) throw new TypeError("Private accessor was defined without a setter");
    if (typeof state === "function" ? receiver !== state || !f : !state.has(receiver)) throw new TypeError("Cannot write private member to an object whose class did not declare it");
    return (kind === "a" ? f.call(receiver, value) : f ? f.value = value : state.set(receiver, value)), value;
};
var __classPrivateFieldGet = (this && this.__classPrivateFieldGet) || function (receiver, state, kind, f) {
    if (kind === "a" && !f) throw new TypeError("Private accessor was defined without a getter");
    if (typeof state === "function" ? receiver !== state || !f : !state.has(receiver)) throw new TypeError("Cannot read private member from an object whose class did not declare it");
    return kind === "m" ? f : kind === "a" ? f.call(receiver) : f ? f.value : state.get(receiver);
};
var _B_a;
class A {
    constructor(a) {
        this.a = a;
    }
    inc() {
        this.a++;
    }
}
class B {
    constructor(x) {
        _B_a.set(this, void 0);
        __classPrivateFieldSet(this, _B_a, x, "f");
    }
    inc() {
        var _a;
        __classPrivateFieldSet(this, _B_a, (_a = __classPrivateFieldGet(this, _B_a, "f"), _a++, _a), "f");
    }
}
_B_a = new WeakMap();
```

I used [`tsconfig.json` from Hono's project](https://github.com/honojs/hono/blob/main/tsconfig.json).

I think this result makes sense.